Code description
This code defines a class TicTacToeGUI that inherits from QMainWindow. In the constructor, it creates the players, the game board (which is a vertical layout containing nine QPushButton widgets), and the status label. It also sets up the main layout and sets the central widget. Finally, it starts the game by setting self.current_player to self.x_player and calling self.update_board().

The update_board() method updates the text of the squares to match the game board and sets the text of the status label depending on the state of the game.

The square_clicked() method is called when a square is clicked. It makes a move in the game using the current player's letter and updates the board. If there is a winner or a tie, it disables all the squares.

The end_game() method disables all the squares.

To run the game, simply execute the `